modify

+++ datasets/dataset.py
@@ -120,7 +120,7 @@ class Dataset(data.Dataset):
                     mask_new[mask == value] = self.label_map[label]
             mask = mask_new
             image, mask, _ = self.transforms(image, mask)
-            return image, mask, idx
+            return image, mask.astype(np.int64), idx
         elif self.config.train_on == "edges":
             image, mask, idx, labels = (
                 Dataset.dataset["images"][dataset_index],
@@ -129,7 +129,7 @@ class Dataset(data.Dataset):
                 Dataset.dataset["labels"][dataset_index],
             )
             image, mask, _ = self.transforms(image, mask)
-            return image, mask, idx
+            return image, mask.astype(np.int64), idx
         elif self.config.train_on == "distance":
             image, shape_distance, neighbor_distance, idx, labels = (
                 Dataset.dataset["images"][dataset_index],
